23.10.05 - 연차 수정폼 및 삭제 구현

+++ src/main/java/com/project/ownote/annualData/dto/AnnData.java
@@ -18,8 +18,8 @@ public class AnnData {
 
   private int  anndata_date;
     private int anndata_add;
-    private int anndata_used;
-    private int anndata_extra;
+    private double anndata_used;
+    private double anndata_extra;
     private int emp_id;
     private int annual_id;
 }
